+++ mot-api/module/StatisticsApi/src/TesterQualityInformation/TesterPerformance/TesterAtSite/Repository/TesterAtSiteSingleGroupStatisticsRepository.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Dvsa\Mot\Api\StatisticsApi\TesterQualityInformation\TesterPerformance\TesterAtSite\Repository;
 
 use Dvsa\Mot\Api\StatisticsApi\TesterQualityInformation\TesterPerformance\Common\Repository\SingleGroupStatisticsRepository;
@@ -8,17 +9,17 @@ use DvsaCommon\Factory\AutoWire\AutoWireableInterface;
 
 class TesterAtSiteSingleGroupStatisticsRepository extends SingleGroupStatisticsRepository implements AutoWireableInterface
 {
-    const PARAM_SITE_ID = "vtsId";
-    const PARAM_TESTER_ID = "testerId";
+    const PARAM_SITE_ID = 'vtsId';
+    const PARAM_TESTER_ID = 'testerId';
 
     public function get($siteId, $testerId, $groupCode, $year, $month)
     {
         return $this->getByParams([
-            self::PARAM_SITE_ID    => $siteId,
-            self::PARAM_TESTER_ID  => $testerId,
+            self::PARAM_SITE_ID => $siteId,
+            self::PARAM_TESTER_ID => $testerId,
             self::PARAM_GROUP_CODE => $groupCode,
-            self::PARAM_YEAR       => $year,
-            self::PARAM_MONTH      => $month,
+            self::PARAM_YEAR => $year,
+            self::PARAM_MONTH => $month,
         ]);
     }
 
@@ -38,11 +39,11 @@ class TesterAtSiteSingleGroupStatisticsRepository extends SingleGroupStatisticsR
     {
         $dbResult = new TesterAtSitePerformanceResult();
         $dbResult
-            ->setTotalTime((double)$row['totalTime'])
-            ->setFailedCount((int)$row['failedCount'])
-            ->setAverageVehicleAgeInMonths((float)$row['averageVehicleAgeInMonths'])
+            ->setTotalTime((float) $row['totalTime'])
+            ->setFailedCount((int) $row['failedCount'])
+            ->setAverageVehicleAgeInMonths((float) $row['averageVehicleAgeInMonths'])
             ->setIsAverageVehicleAgeAvailable(!is_null($row['averageVehicleAgeInMonths']))
-            ->setTotalCount((int)$row ['totalCount'])
+            ->setTotalCount((int) $row ['totalCount'])
             ->setSiteName($row['siteName'])
         ;
